ingestion/requirements.txt
# Core DB
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# RSS Parsing
feedparser==6.0.11

# HTTP Requests (for future APIs like Reddit, YouTube, arXiv)
requests==2.31.0

# Date handling (optional but useful)
python-dateutil==2.9.0.post0

# Environment variables
python-dotenv==1.0.1